Assignment/Module 17) Rest Framework/RestProject/requirements.txt
django>=4.2,<5.0
djangorestframework
requests
django-cors-headers
django-allauth
twilio
sendgrid
stripe
python-dotenv
